+++ src/main/java/es/gob/minetad/doctopic/CleanZeroEpsylonIndex.java
@@ -44,8 +44,8 @@ public class CleanZeroEpsylonIndex implements DocTopicsIndex{
     }
 
     @Override
-    public Double getPrecision() {
-        return Double.valueOf(precision);
+    public Integer getPrecision() {
+        return Float.valueOf(precision).intValue();
     }
 
     @Override
