+++ src/main/java/es/gob/minetad/doctopic/CleanZeroEpsylonTVMIndex.java
@@ -74,8 +74,8 @@ public class CleanZeroEpsylonTVMIndex implements DocTopicsIndex{
     }
 
     @Override
-    public Double getPrecision() {
-        return Double.valueOf(precision);
+    public Integer getPrecision() {
+        return Float.valueOf(precision).intValue();
     }
 
     @Override
