+++ src/main/java/es/gob/minetad/doctopic/SimGraphArrayIndex.java
@@ -80,8 +80,8 @@ public class SimGraphArrayIndex implements DocTopicsIndex{
     }
 
     @Override
-    public Double getPrecision() {
-        return Double.valueOf(precision);
+    public Integer getPrecision() {
+        return Float.valueOf(precision).intValue();
     }
 
     @Override
